requirements.txt
Django==5.0.1
pillow==10.2.0
reportlab==4.0.9
python-dateutil==2.8.2
gunicorn==21.2.0
whitenoise==6.6.0
openpyxl==3.1.2
django-allauth==65.14.0
